Feature/code coverage

+++ backend/src/main/java/com/swarmer/finance/models/TransactionTypeConverter.java
@@ -6,15 +6,15 @@ import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
-public class TransactionTypeConverter  implements AttributeConverter<TransactionType, Integer> {
+public class TransactionTypeConverter  implements AttributeConverter<TransactionType, Long> {
 
     @Override
-    public Integer convertToDatabaseColumn(TransactionType attribute) {
+    public Long convertToDatabaseColumn(TransactionType attribute) {
         return attribute == null ? null : attribute.getValue();
     }
 
     @Override
-    public TransactionType convertToEntityAttribute(Integer dbData) {
+    public TransactionType convertToEntityAttribute(Long dbData) {
         if (dbData == null) {
             return null;
         }
